add afterfunc usage example
